Actualizacion estructura
