--- mercadopago/accessToken.go
@@ -1,6 +1,0 @@
-package mercadopago
-
-func accessToken() string {
-
-	return "TEST-7981108178642173-052023-cd49dbc0e54c62bf03db3bbfe9b13937-1104296870"
-}
