+++ mercadopago/webhook.go
@@ -1,13 +1,22 @@
 package mercadopago
 
 import (
+	"fmt"
+
 	"github.com/gin-gonic/gin"
 )
 
-func get(c *gin.Context) {
+var request map[string]interface{}
+
+func Webhook(c *gin.Context) {
+
+	c.BindJSON(&request)
+
+	fmt.Println(request)
 
 	c.JSON(200, gin.H{
 
-		"message": "HOLA",
+		"message":   "Webhook Recibido",
+		"respuesta": request,
 	})
 }
