+++ routes/routes.go
@@ -1,6 +1,11 @@
 package routes
 
 import (
+	"backend-facturacion/handlers"
+	"backend-facturacion/mercadopago"
+	"time"
+
+	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 )
 
@@ -9,7 +14,7 @@ func SetupRouter() *gin.Engine {
 	r := gin.Default()
 
 	// Middleware CORS
-	/*r.Use(cors.New(cors.Config{
+	r.Use(cors.New(cors.Config{
 		AllowOrigins:     []string{"http://localhost:3050"},
 		AllowMethods:     []string{"GET"},
 		AllowHeaders:     []string{"Origin", "Content-Type"},
@@ -17,16 +22,10 @@ func SetupRouter() *gin.Engine {
 		MaxAge:           12 * time.Hour,
 	}))
 
-	// Endpoint público
+	//Endpoints
+	r.POST("/API/v1/webhook", mercadopago.Webhook)
+	r.POST("/API/v1/payment", mercadopago.Payment)
 	r.GET("/api/cotizacion/:id", handlers.GetCotizacionByID)
-
-	r.GET("/", func(c *gin.Context) {
-		c.JSON(200, gin.H{
-			"message": "pong",
-		})
-	})*/
-
-	// Aquí se pueden agregar más rutas
 
 	return r
 }
